feat: clean up code

See merge request GuiParpineli/clearContWebApp!3

+++ src/main/java/com/clearcont/clearcontapp/model/ConciliacaoContabil.java
@@ -6,7 +6,6 @@ import jakarta.persistence.GenerationType;
 import jakarta.persistence.Id;
 import lombok.Getter;
 import lombok.Setter;
-import org.atmosphere.config.service.Get;
 
 import java.time.LocalDateTime;
 
@@ -16,7 +15,7 @@ import java.time.LocalDateTime;
 public class ConciliacaoContabil {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer id;
+    private Long id;
     private String status;
     private LocalDateTime data;
     private String classificacao;
