+++ src/main/java/com/clearcont/clearcontapp/model/EmpresaGroup.java
@@ -1,15 +1,11 @@
 package com.clearcont.clearcontapp.model;
 
 import jakarta.persistence.*;
-import jakarta.validation.constraints.NotNull;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
-import org.springframework.security.core.GrantedAuthority;
-import org.springframework.security.core.userdetails.UserDetails;
 
-import java.util.Collection;
 import java.util.List;
 
 @Getter
@@ -20,7 +16,7 @@ import java.util.List;
 public class EmpresaGroup {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer id;
+    private Long id;
     @OneToMany(fetch = FetchType.EAGER)
     public List<Empresa> empresas;
 }
