+++ src/main/java/com/clearcont/clearcontapp/repository/ControleRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface ControleRepository extends JpaRepository<Controle, Integer> {
+public interface ControleRepository extends JpaRepository<Controle, Long> {
 }
